// src/app/modules/ChampPermanentReplacements/Interfaces.tsx
export interface ChampsData {
    id:                                 number;
    name:                               string;
    email:                              string;
    phone:                              string;
    flatno:                             string;
    locality:                           string;
    address:                            string;
    latitude:                           string;
    longitude:                          string;
    city:                               number;
    pincode:                            number;
    vehicleid:                          number;
    frequencyid:                        number;
    jobsiteid:                          number;
    customerid:                         number;
    cleanerid:                          number;
    uplevelcleanerid:                   number;
    orderid:                            number;
    cleanerprice:                       number;
    packageprice:                       number;
    discountprice:                      number;
    discountfor:                        string;
    gst:                                number;
    total:                              number;
    startdate:                          Date;
    enddate:                            Date;
    packageid:                          number;
    packagedays:                        number;
    timeslot:                           string;
    fullcleaningday:                    string;
    paymentdate:                        Date;
    transactionid:                      string;
    paymenttype:                        string;
    status:                             number;
    isactive:                           number;
    payment_mode:                       string;
    order_type:                         string;
    razorpay_order_id:                  string;
    razorpay_autopay_initated_order_id: string;
    razorpay_signature:                 string;
    payment_method_id:                  number;
    parking_no:                         string;
    extra_amount:                       number;
    is_autopay:                         number;
    razorpay_token:                     string;
    razorpay_status:                    string;
    parent_order_id:                    number;
    inactive_date:                      string;
    reason_id:                          number;
    fine_amount:                        number;
    fine_for:                           string;
    address_id:                         number;
    ct_offer_id:                        number;
    refund_amount:                      number;
    refund_id:                          string;
    refund_initiated_date:              string;
    is_settled:                         number;
    settled_for_subscription_id:        number;
    is_paused:                          number;
    months:                             number;
    razorpay_payment_link_id:           string;
    invoice_status:                     number;
    createdAt:                          number;
    updatedAt:                          number;
    jobsitename:                          string;

}